\name{seqMeta-package}
\alias{seqMeta-package}
\alias{seqMeta}
\docType{package}
\title{
\packageTitle{seqMeta}
}
\description{
\packageDescription{seqMeta}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{seqMeta}
\packageIndices{seqMeta}
%% ~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{seqMeta}

Maintainer: \packageMaintainer{seqMeta}
}
\keyword{ package }
