% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqMeta.R
\docType{package}
\name{seqMeta}
\alias{seqMeta}
\alias{seqMeta-package}
\title{seqMeta: Meta-Analysis of Region-Based Tests of Rare DNA Variants}
\description{
Computes necessary information to meta analyze region-based tests for rare genetic variants (e.g. SKAT, T1) in individual studies, and performs meta analysis.
}
\details{
To learn more about seqMeta, start with the vignettes:
\code{browseVignettes(package = "seqMeta")}
}

